+++ CotEditor/Sources/DefaultKeys.swift
@@ -8,7 +8,7 @@
 //
 //  ---------------------------------------------------------------------------
 //
-//  © 2016-2023 1024jp
+//  © 2016-2024 1024jp
 //
 //  Licensed under the Apache License, Version 2.0 (the "License");
 //  you may not use this file except in compliance with the License.
@@ -157,7 +157,7 @@ extension DefaultKeys {
     static let selectedInspectorPaneIndex = DefaultKey<Int>("selectedInspectorPaneIndex")
     static let splitViewVertical = DefaultKey<Bool>("splitViewVertical")
     static let consoleFontSize = DefaultKey<CGFloat>("consoleFontSize")
-    static let outlineViewFontSize = DefaultKey<CGFloat>("outlineViewFontSize")
+    static let outlineViewFontSize = DefaultKey<Double>("outlineViewFontSize")
     static let findResultViewFontSize = DefaultKey<Double>("findResultViewFontSize")
     static let selectedMultipleReplaceSettingName = DefaultKey<String?>("selectedMultipleReplaceSettingName")
     static let unicodeHistory = DefaultKey<[String]>("unicodeHistory")